+++ src/lexer/model/definition.ts
@@ -1,8 +1,8 @@
 import type { Action } from "../action";
 
 /** Apply `action` and try to yield a token with `kind`. */
-export type Definition<Data, ErrorType, Kinds extends string, ActionState> = {
+export type Definition<Kinds extends string, Data, ActionState, ErrorType> = {
   /** Target token kind. Empty string if anonymous. */
   kind: Kinds;
-  action: Action<Data, ErrorType, ActionState>;
+  action: Action<Data, ActionState, ErrorType>;
 };